Code layout refactored to separate ground and angular systems to improve readability.
