+++ ctapipe/coordinates/__init__.py
@@ -2,4 +2,5 @@
 """
 Coordinates.
 """
-from .frames import *
+from .angular_frames import *
+from .ground_frames import *
